+++ SharkScope/Model/Tournaments/TournamentsRequest.swift
@@ -1,0 +1,9 @@
+//
+//  TournamentsQuery.swift
+//  SharkScope
+//
+//  Created by Geri Borbás on 2020. 01. 19..
+//  Copyright © 2020. Geri Borbás. All rights reserved.
+//
+
+import Foundation